updated cpp to latest version

+++ cpp/token2.c
@@ -6,6 +6,6 @@
 #ifdef __GNUC__
 __inline
 #endif
-static unsigned int hash2 P((register const char *, register unsigned int));
+static unsigned int hash2 P((register const char *, register size_t));
 
 #include "token2.h"
